erajs-backend/erajs/requirements.txt
colorama
cryptography
email-validator
fastapi
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pyyaml
uvicorn[standard]
semver